requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.5
numpy==1.26.4
pandas==2.2.1
pytest==7.4.3
sqlalchemy==2.0.23
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
httpx==0.25.2
websockets==12.0
aiohttp==3.9.1
plotly==5.18.0
dash==2.14.1
streamlit==1.29.0
ta==0.11.0
beautifulsoup4==4.12.2
transformers==4.38.2
yfinance==0.2.37
scikit-learn==1.4.1.post1
shap==0.44.1
python-dateutil==2.8.2
lime==0.2.0.1
eli5==0.13.0
matplotlib==3.8.2
nltk==3.8.1 